+++ src/mod.ts
@@ -2,10 +2,8 @@
  * @module @run-z/project-config
  */
 export * from './gitignore/mod.js';
+export * from './package/mod.js';
 export * from './project-config.js';
-export * from './project-entry.js';
-export * from './project-export.js';
 export * from './project-output.js';
-export * from './project-package.js';
 export * from './project-tools-init.js';
 export * from './typescript/mod.js';
